requirements.txt
PySide6>=6.6
python-vlc>=3.0.20123
requests>=2.31
packaging>=23.0
